ci: Improve CI - remove flox, keep it only for local development (#162)
## Summary

Currently when claude is running the CI there is no option for it to
execute the commands since they are in a flox environment.

Sure, I can add an instruction to cluade to run `flox activate --
<command>`, but this could get unreliable.

I also notice there is a flox activate action, but this only allow me to
run the command within the action.

To keep it simple, I'll duplicate the CI dependencies from the flox
environment and this way there would also be better caching, because
those dependencies live directly on the ephemeral runner.

Also leveraging the opportunity to upgrade golang to 1.25

I still love flox, but I'll keep it only for the local sandbox
environment.

Closes #161

---------

Signed-off-by: Eden Reich <eden.reich@gmail.com>