refactor: Improve the extensibility of the CLI and support multiple container runtimes
This will help me in the future to add podman or containerd should I need to. The goal is container native CLI where AI is isolated in a sandbox and the inference is consumed from a service.

Starting with docker as a default runtime, I still need to test the this CLI also works well without any container runtime - standalone gateway binary, regression tests are needed.

Signed-off-by: Eden Reich <eden.reich@gmail.com>